+++ app/src/main/java/com/mindpalace/app/domain/usecase/GetFragmentsByLastOpenedUseCase.kt
@@ -4,7 +4,7 @@ import com.mindpalace.app.domain.model.MindFragmentSummary
 import com.mindpalace.app.domain.repository.MindFragmentRepository
 
 class GetFragmentsByLastOpenedUseCase (private val repository: MindFragmentRepository) {
-    suspend operator fun invoke(userId: String, limit: Int = 6): Result<List<MindFragmentSummary>> {
+    suspend operator fun invoke(userId: String, limit: Long = 6): Result<List<MindFragmentSummary>> {
         return repository.getFragmentsByLastOpened(userId, limit)
     }
 }
